Fix #17:
The animation queueing code to support staggered animations wasn't correctly firing all ngAnimate animation complete callbacks.

This caused DOM nodes to be left in the DOM after ngEnter or ngLeave transitions had operated on more than one element at a time.

This happened because the Velocity animation complete callback was not calling the complete function of the whole animation queue, but just the last item in the queue.

Fix by making sure the completion function that runs for the Velocity animation triggers the queue completeion function.